pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "telnyx-transcribe"
version = "2.0.0"
description = "🎙️ Automated call-and-transcribe solution using Telnyx API and OpenAI Whisper"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Yiğit Konur", email = "yigit@konur.dev" },
]
keywords = [
    "telnyx",
    "openai",
    "whisper",
    "transcription",
    "call",
    "voice",
    "speech-to-text",
    "telephony",
    "automation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Telephony",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "telnyx>=2.0.0",
    "openai>=1.0.0",
    "flask>=3.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
telnyx-transcribe = "telnyx_transcribe.cli:run"
tt = "telnyx_transcribe.cli:run"

[project.urls]
Homepage = "https://github.com/yigitkonur/telnyx-transcribe"
Documentation = "https://github.com/yigitkonur/telnyx-transcribe#readme"
Repository = "https://github.com/yigitkonur/telnyx-transcribe.git"
Issues = "https://github.com/yigitkonur/telnyx-transcribe/issues"
Changelog = "https://github.com/yigitkonur/telnyx-transcribe/releases"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/telnyx_transcribe"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip strict
]

[tool.ruff.isort]
known-first-party = ["telnyx_transcribe"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["telnyx.*", "flask.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --cov=telnyx_transcribe --cov-report=term-missing"

[tool.coverage.run]
source = ["src/telnyx_transcribe"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
